feat(portfolio): move portfolio nav into shared PageHeader
Replaces the bespoke PrototypeNav with a HeaderNav embedded in
PageHeader, used by both /app and the (portfolio) route group. The
new tabs are Explore (back to /app, MapTrifoldIcon), Inbox, Reports,
Dashboards, each with an active-state pill that respects pathname.

Reverts the sidebar Portfolio section (the header is the canonical
nav now).
