+++ app/(portfolio)/layout.tsx
@@ -2,7 +2,7 @@
 
 import { Box } from "@chakra-ui/react";
 import { useAuthGuard } from "@/app/hooks/useAuthGuard";
-import PrototypeNav from "@/app/components/portfolio/PrototypeNav";
+import PageHeader from "@/app/components/PageHeader";
 
 export default function PortfolioLayout({
   children,
@@ -14,7 +14,7 @@ export default function PortfolioLayout({
 
   return (
     <Box display="flex" flexDir="column" minH="100vh" bg="bg.subtle">
-      <PrototypeNav />
+      <PageHeader />
       <Box flex="1" minH={0}>
         {children}
       </Box>
